refactor the /single/precomputed request implementation
