+++ src/server/main.py
@@ -1,28 +1,16 @@
-from typing import Optional, List, Dict
-from pydantic import BaseModel
-
 from fastapi import FastAPI
 import pandas as pd
 
 import src.server.single.precomputed as sp
 
-class Person(BaseModel):
-    id: str
-
-class Response(BaseModel):
-    speaking_time: Dict[str, List[List]]
-    relative_diff: Dict[str, List[List]]
-    unanchored: Dict[str, List[List]]
-    wpm: Dict[str, List[List]]
-
 app = FastAPI()
 
 
-@app.post("/single/precomputed", response_model=Response)
-async def single_precomputed(person: Person):
+@app.post("/single/precomputed", response_model=sp.Response)
+async def single_precomputed(request: sp.Request):
     path = "samples/sample_statistics_output/precomputed/all.txt"
     all_df = pd.read_csv(path)
-    speaker_df = all_df.loc[all_df["speaker"] == person.id]
+    speaker_df = all_df.loc[all_df["speaker"] == request.MoP]
 
     response = {
         "speaking_time": dict(),
@@ -33,5 +21,5 @@ async def single_precomputed(person: Person):
 
     speaker_df.apply(lambda row: sp.update_response(response, row) , axis=1)
 
-    return Response(**response)
+    return sp.Response(**response)
